# public/swagger.json
{
	"components": {
...
		"schemas": {
			"VerificationMethod": {
				"properties": {
					"id": {
						"type": "string"
					},
					"controller": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"Ed25519VerificationKey2018"
						],
						"nullable": false
					},
					"publicKeyMultibase": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"controller",
					"type",
					"publicKeyMultibase"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Service": {
				"properties": {
					"id": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"LinkedDomains",
							"DIDCommMessaging"
						]
					},
					"serviceEndpoint": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"type",
					"serviceEndpoint"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DidDocument": {
				"properties": {
					"@context": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"controller": {
						"type": "string"
					},
					"verificationMethod": {
						"items": {
							"$ref": "#/components/schemas/VerificationMethod"
						},
						"type": "array"
					},
					"authentication": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"assertionMethod": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"keyAgreement": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"capabilityInvocation": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"capabilityDelegation": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"service": {
						"items": {
							"$ref": "#/components/schemas/Service"
						},
						"type": "array"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"@context",
					"verificationMethod",
					"authentication",
					"assertionMethod",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidateErrorJSON": {
				"properties": {
					"message": {
						"type": "string",
						"enum": [
							"Validation failed"
						],
						"nullable": false
					},
					"details": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					}
				},
				"required": [
					"message",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IDidDocumentRegisterPayload": {
				"properties": {
					"publicKeyMultibase": {
						"type": "string"
					}
				},
				"required": [
					"publicKeyMultibase"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IDidOwnershipClaimPayload": {
				"properties": {
					"privateKeyMultibase": {
						"type": "string"
					}
				},
				"required": [
					"privateKeyMultibase"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IDidOwnershipRegisterPayload": {
				"properties": {
					"privateKeyMultibase": {
						"type": "string"
					}
				},
				"required": [
					"privateKeyMultibase"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IServiceRegisterPayload": {
				"properties": {
					"service": {
						"$ref": "#/components/schemas/Service"
					}
				},
				"required": [
					"service"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IServiceUpdateBody": {
				"properties": {
					"type": {
						"type": "string",
						"enum": [
							"LinkedDomains",
							"DIDCommMessaging"
						]
					},
					"serviceEndpoint": {
						"type": "string"
					}
				},
				"required": [
					"type",
					"serviceEndpoint"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IServiceUpdatePayload": {
				"properties": {
					"service": {
						"$ref": "#/components/schemas/IServiceUpdateBody"
					}
				},
				"required": [
					"service"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"SignedRequestHeader": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			},
			"DigestHeader": {
				"type": "apiKey",
				"name": "Digest",
				"in": "header"
			},
			"ExpiresHeader": {
				"type": "apiKey",
				"name": "Expires",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "hedera-appnet",
		"version": "1.0.0",
		"description": "This is a hedera appnet API application",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Meeco",
			"url": "https://meeco.me",
			"email": "‍support@meeco.me"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/did": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "DidDocument",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidDocument"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Register a new DID document. User provides public key that is going to be added as a delegate key that allows user to modify created DID document later.",
				"summary": "Register a new DID Document.",
				"tags": [
					"Document"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IDidDocumentRegisterPayload"
							}
						}
					}
				}
			}
		},
		"/did/{did}": {
			"get": {
				"operationId": "Resolve",
				"responses": {
					"200": {
						"description": "DidDocument",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidDocument"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Resolve DID Document",
				"summary": "Resolve DID Document",
				"tags": [
					"Document"
				],
				"security": [],
				"parameters": [
					{
						"description": "A percent-escaped DID Identifier as defined in DID specification <br /> <br />\nExample: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231",
						"in": "path",
						"name": "did",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "Revoke",
				"responses": {
					"204": {
						"description": "No content"
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Permanently remove DID Document from Appnet registry. In addition to that, new messages will be written to the DID topic stating that document has been removed.",
				"summary": "Remove DID Document from registry",
				"tags": [
					"Document"
				],
				"security": [
					{
						"SignedRequestHeader": [],
						"DigestHeader": [],
						"ExpiresHeader": []
					}
				],
				"parameters": [
					{
						"description": "A percent-escaped DID Identifier as defined in DID specification<br /> <br />\nExample: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231",
						"in": "path",
						"name": "did",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/did/{did}/claim": {
			"post": {
				"operationId": "Claim",
				"responses": {
					"200": {
						"description": "DidDocument",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidDocument"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Claim DID Document ownership back from the AppNet. Changes DID root key to the one provided via parameters. DID controller remains the same.",
				"summary": "Claim DID Document ownership back from the AppNet",
				"tags": [
					"Ownership"
				],
				"security": [
					{
						"SignedRequestHeader": [],
						"DigestHeader": [],
						"ExpiresHeader": []
					}
				],
				"parameters": [
					{
						"description": "A percent-escaped DID Identifier as defined in DID specification <br /> <br />\nExample: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231",
						"in": "path",
						"name": "did",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IDidOwnershipClaimPayload"
							}
						}
					}
				}
			}
		},
		"/did/{did}/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "void",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidDocument"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Register an existing DID Document with AppNet. Gives away control of the document to the AppNet component.",
				"summary": "Register an existing DID Document with AppNet",
				"tags": [
					"Ownership"
				],
				"security": [],
				"parameters": [
					{
						"description": "A percent-escaped DID Identifier as defined in DID specification<br /> <br />\nExample: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231",
						"in": "path",
						"name": "did",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IDidOwnershipRegisterPayload"
							}
						}
					}
				}
			}
		},
		"/did/{did}/services": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidDocument"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Register a new service to the DID Document",
				"summary": "Register a new service to the DID Document",
				"tags": [
					"Service"
				],
				"security": [
					{
						"SignedRequestHeader": [],
						"DigestHeader": [],
						"ExpiresHeader": []
					}
				],
				"parameters": [
					{
						"description": "A percent-escaped DID Identifier as defined in DID specification <br /> <br />\nExample: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231",
						"in": "path",
						"name": "did",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Register service payload",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IServiceRegisterPayload"
							}
						}
					}
				}
			}
		},
		"/did/{did}/services/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "DidDocument",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidDocument"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Update service information on the DID Document",
				"summary": "Update service information on the DID Document",
				"tags": [
					"Service"
				],
				"security": [
					{
						"SignedRequestHeader": [],
						"DigestHeader": [],
						"ExpiresHeader": []
					}
				],
				"parameters": [
					{
						"description": "A percent-escaped DID Identifier as defined in DID specification <br /> <br />\nExample: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231",
						"in": "path",
						"name": "did",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "A percent-escaped Service ID string <br /> <br />\nExample: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231%23service-1",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Update service payload",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IServiceUpdatePayload"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Revoke",
				"responses": {
					"200": {
						"description": "DidDocument",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DidDocument"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateErrorJSON"
								}
							}
						}
					}
				},
				"description": "Remove service information from the DID Document",
				"summary": "Remove service information from the DID Document",
				"tags": [
					"Service"
				],
				"security": [
					{
						"SignedRequestHeader": [],
						"DigestHeader": [],
						"ExpiresHeader": []
					}
				],
				"parameters": [
					{
						"description": "A percent-escaped DID Identifier as defined in DID specification <br /> <br />\nExample: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231",
						"in": "path",
						"name": "did",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "A percent-escaped Service ID string <br /> <br />\nExample: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231%23service-1",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}