# src/controllers/did-service.controller.ts
import {
  Body,
  Controller,
  Delete,
  Path,
  Post,
  Put,
  Response,
  Route,
  Security,
  SuccessResponse,
  Tags,
} from "tsoa";
import {
  DidDocument,
  IServiceRegisterPayload,
  IServiceUpdatePayload,
  ValidateErrorJSON,
} from "../models";
import { registerService, revokeService, updateService } from "../services";

@Route("did")
@Tags("Service")
export class DidServiceController extends Controller {
  /**
   * Register a new service to the DID Document
   * @summary Register a new service to the DID Document
   * @param did A percent-escaped DID Identifier as defined in DID specification <br /> <br />
   * Example: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231
   * @param body Register service payload
   * @returns DidDocument
   */
  @Response<ValidateErrorJSON>(422, "Validation Failed")
  @SuccessResponse(201, "Created")
  @Security({ SignedRequestHeader: [], DigestHeader: [], ExpiresHeader: [] })
  @Post("/{did}/services")
  public async register(
    @Path() did: string,
    @Body() body: IServiceRegisterPayload
  ): Promise<DidDocument> {
    return registerService(decodeURIComponent(did), body);
  }

  /**
   * Update service information on the DID Document
   * @summary Update service information on the DID Document
   * @param did A percent-escaped DID Identifier as defined in DID specification <br /> <br />
   * Example: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231
   * @param id A percent-escaped Service ID string <br /> <br />
   * Example: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231%23service-1
   * @param body Update service payload
   * @returns DidDocument
   */
  @Response<ValidateErrorJSON>(422, "Validation Failed")
  @Security({ SignedRequestHeader: [], DigestHeader: [], ExpiresHeader: [] })
  @Put("/{did}/services/{id}")
  public async update(
    @Path() did: string,
    @Path() id: string,
    @Body() body: IServiceUpdatePayload
  ): Promise<DidDocument> {
    return updateService(decodeURIComponent(did), decodeURIComponent(id), body);
  }

  /**
   * Remove service information from the DID Document
   * @summary Remove service information from the DID Document
   * @param did A percent-escaped DID Identifier as defined in DID specification <br /> <br />
   * Example: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231
   * @param id A percent-escaped Service ID string <br /> <br />
   * Example: did%3Ahedera%3Atestnet%3Az6MkubW6fwkWSA97RbKs17MtLgWGHBtShQygUc5SeHueFCaG_0.0.29656231%23service-1
   * @returns DidDocument
   */
  @Response<ValidateErrorJSON>(422, "Validation Failed")
  @Security({ SignedRequestHeader: [], DigestHeader: [], ExpiresHeader: [] })
  @Delete("/{did}/services/{id}")
  public async revoke(
    @Path() did: string,
    @Path() id: string
  ): Promise<DidDocument> {
    return revokeService(decodeURIComponent(did), decodeURIComponent(id));
  }
}